add code generator
This takes the existing API docs from production
and generates an object with methods that do a
one-to-one map with API methods in production.

In the future, we should take advantage of the input/output
metadata to flesh out the responses as actual objects they
can interact with, but for now this works.